Add git aliases for fixup and checking out files
git-cof: If you are in the root of a repo and use this command you will
get a typeahead that allows you to select a modified file to check out.

git-fixup: This is an alias for `git commit --fixup`. It defaults to
fixing up against the previous commit without you having to pass the
hash of that commit.