Add pprof binaries
The `pprof` binary is the newer of the two, implemented in Go.

The `pprof-nongo` binary is an earlier implementation of the same thing.
However, it is still required in some cases because it supports
operations (and pprof data formats) that are not supported by the more
current pprof. One example of this is the Twitter CPU profiling tool
built into the Finagle admin dashboard.