Install selecta on Debian
Adds a Debian-specific task to install the selecta binary into the ~/bin
directory as a symlink to the selecta binary in the selecta repo (included
as a submodule in this commit). If this goes well, it could also replace the
selecta install on Darwin which is currently managed with Homebrew.