Add .inputrc file and set revert-all-at-newline
I've long had vi mode set in my bashrc but moving it to be configured on
inputrc makes sense to be more broadly applicable to other programs. In
addition, prevents history mangling due to not putting commands that
were derived from history but not run into the history. Thanks to
https://twobithistory.org/2019/08/22/readline.html for that tidbit.