Configure sbt server and vim ALE to use it
This commit adds support for ALE to use the LSP server that is built in
to sbt for Scala. It compiles the project on saves and will highlight if
imports are missing or syntax is incorrect. Presumably it does other
things as well that I haven't quite worked out yet.

The ale-scala docs [1] noted that the suggested way to get the sbt
server to work with ALE is to change the server from using a unix socket
to using tcp. That is what the configuration in global.sbt does. The
.vimrc change disables other scala linters/compilers for now as (I
suspect scalac, but could have been another) one of them always
complained about undefined objects at package boundaries. The sbtserver
plugin does not have this problem.

I had to implement some additional ansible plumbing for this to work.
Specifying .sbt as the directory to hard link resulted in overwriting
the rest of the .sbt directory (which is undesired) and for arbitrary
.sbt files to be written into the repository. There was previously not a
way to put a file into a nested directory where the directory may or may
not have existed. My change introduces a second list of files that have
a path (underneath the home directory) that is first created if it does
not exist and then a file name to link into that path.

[1]: https://github.com/dense-analysis/ale/blob/4afbf2f25dc0ce86b118261b0cfb904c80ae6ba0/doc/ale-scala.txt#L40