Add srcery theme for vim and iterm
iterm file comes from commit 4fff588 of
https://github.com/srcery-colors/srcery-terminal

Theme not yet enabled by default in vim.

Removes lexima as I found it equally helpful and annoying depending on
the circumstance. I may be able to tune it to turn it off when editing
vs adding new code but for now just removing entirely.