Configuration for 3 new home servers
- two optiplex 3060
- one optiplex 3080

Defines morph configuration for deploying config to them as well as
barebones config for the system and creates a user account.

Still need to refactor config for my desktop to share home manager
config across these servers for my user account and also begin to set up
services on these.