Add service for replicating zfs snapshots to NAS
First pass on using `zfs-send` to send snapshots to a remote host.

Future work on this could make it a full module with some
configurability:

- configurable remote user/host and destination path
- configurable scheduled run time
- which snapshots to filter (daily auto-snapshots only for now)

Additionally there's some future work that could be done such as:

- adding jitter so all hosts don't send snapshots at exactly the same
  time (using e.g. RandomizedDelaySec and maybe FixedRandomDelay)
- consider sending incremental snapshots with full snapshots on some
  frequency to save on remote disk space
- cleaning up old snapshots to save on remote disk space

For now the remote disk space is large enough with the disks on the zfs
machines being small enough that space shouldn't be a problem.