Add transcoding setup
Initially used a custom makemkv package until I realized unstable had
the version I needed. Then pinned a specific unstable hash for the
moment.

I tried to move the unstablePkgs up the function hierarchy a bit, closer
to either top-level for my user or top-level for common but I ran into
issues where I could define unstablePkgs higher up and even include it
as a parameter in the lower functions, but when referencing it I was
told the attribute was not defined.

Not quite clear on how that works so for now I'm leaving it defined in
linuxPackages but I'd like to be able to use the same unstablePkgs
definition in multiple places in the future.